test(server): reorder some test files
