--- server/test/sc/protocol/requests/RoomPacketTest.java
@@ -1,6 +1,0 @@
-package sc.protocol.requests;
-
-import sc.server.network.RealServerTest;
-
-public class RoomPacketTest extends RealServerTest {
-}
